backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiosqlite>=0.20.0
anthropic>=0.40.0
openai>=1.12.0
pydantic>=2.0.0
python-dotenv>=1.0.1
